Fix Spotify percent decoding panic
